+++ cmd/openbee/internal/cli/servicecmd/manager_linux.go
@@ -25,6 +25,7 @@ type systemdTemplateData struct {
 	ExePath    string
 	ConfigPath string
 	LogPath    string
+	WorkingDir string
 	Home       string
 }
 
@@ -65,6 +66,7 @@ func (m linuxManager) Install(ctx context.Context, opts InstallOptions) error {
 		ExePath:    opts.ExePath,
 		ConfigPath: opts.ConfigPath,
 		LogPath:    opts.LogPath,
+		WorkingDir: opts.WorkingDir,
 		Home:       home,
 	})
 	if err != nil {